fixed build errors and warnings.
