+++ src/Dogger/Dogger.csproj
@@ -19,21 +19,21 @@
     <PackageReference Include="Auth0.AuthenticationApi" Version="7.0.9" />
     <PackageReference Include="Auth0.ManagementApi" Version="7.0.9" />
     <PackageReference Include="automapper" Version="10.0.0" />
-    <PackageReference Include="AWSSDK.Core" Version="3.3.107.27" />
-    <PackageReference Include="AWSSDK.ECR" Version="3.3.106.8" />
+    <PackageReference Include="AWSSDK.Core" Version="3.3.107.32" />
+    <PackageReference Include="AWSSDK.ECR" Version="3.3.107.2" />
     <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.3.101" />
-    <PackageReference Include="AWSSDK.IdentityManagement" Version="3.3.106.22" />
-    <PackageReference Include="AWSSDK.Lightsail" Version="3.3.105.21" />
-    <PackageReference Include="awssdk.securitytoken" Version="3.3.105.31" />
+    <PackageReference Include="AWSSDK.IdentityManagement" Version="3.3.106.27" />
+    <PackageReference Include="AWSSDK.Lightsail" Version="3.3.106.4" />
+    <PackageReference Include="awssdk.securitytoken" Version="3.3.105.36" />
     <PackageReference Include="Destructurama.Attributed" Version="2.0.0" />
     <PackageReference Include="docker.dotnet" Version="3.125.2" />
     <PackageReference Include="FluffySpoon.AspNet.NGrok" Version="1.75.0" />
     <PackageReference Include="Flurl.Http" Version="2.4.2" />
     <PackageReference Include="GitHubJwt" Version="0.0.4" />
-    <PackageReference Include="hangfire" Version="1.7.11" />
+    <PackageReference Include="hangfire" Version="1.7.12" />
     <PackageReference Include="Macross.Json.Extensions" Version="1.5.0" />
-    <PackageReference Include="mediatr" Version="8.0.2" />
-    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="8.0.1" />
+    <PackageReference Include="mediatr" Version="8.1.0" />
+    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="8.1.0" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.6" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="3.1.6" />
     <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="3.1.6" />
@@ -57,13 +57,13 @@
     <PackageReference Include="nsubstitute" Version="4.2.2" />
     <PackageReference Include="Octokit" Version="0.48.0" />
     <PackageReference Include="Polly" Version="7.2.1" />
-    <PackageReference Include="Serilog.AspNetCore" Version="3.2.0" />
+    <PackageReference Include="Serilog.AspNetCore" Version="3.4.0" />
     <PackageReference Include="serilog.sinks.elasticsearch" Version="8.2.0" />
     <PackageReference Include="Serilog.Sinks.Slack.Core" Version="0.1.5-beta" />
     <PackageReference Include="SerilogAnalyzer" Version="0.15.0" />
     <PackageReference Include="Slack.Webhooks" Version="1.1.4" />
     <PackageReference Include="SSH.NET" Version="2016.1.0" />
-    <PackageReference Include="Stripe.net" Version="37.24.0" />
+    <PackageReference Include="Stripe.net" Version="37.27.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.5.1" />
     <PackageReference Include="YamlDotNet" Version="8.1.2" />
   </ItemGroup>
